03_conversational_chat/requirements.txt
# Production Conversational AI Requirements
# Multi-platform AI integration with production error handling

# Web framework
Flask>=2.0.0
flask-cors>=4.0.0

# AI Platform APIs
google-generativeai>=0.3.0  # Gemini Pro
openai>=1.0.0               # OpenAI GPT-4
google-cloud-aiplatform>=1.35.0  # Invertex AI (Vertex AI)

# Core ML and NLP
numpy>=1.21.0
scikit-learn>=1.3.0
nltk>=3.8.0
transformers>=4.30.0

# Text processing and search
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
tiktoken>=0.5.0
rapidfuzz>=3.0.0

# Image processing (for visual part identification)
Pillow>=10.0.0
opencv-python>=4.8.0

# Data handling
pandas>=2.0.0
requests>=2.31.0
aiohttp>=3.8.0

# Performance and caching
redis>=4.5.0
asyncio  # Built-in

# Production features
python-dotenv>=1.0.0
pydantic>=2.0.0
tenacity>=8.2.0

# Logging and monitoring
loguru>=0.7.0
prometheus-client>=0.17.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
